Split ios/android MenuPicker into separate files
